more inetdiag request to inetdiag pkg
