+++ collector/socket-monitor.go
@@ -9,11 +9,9 @@ import "C"
 */
 
 import (
-	"errors"
 	"log"
 	"syscall"
 	"time"
-	"unsafe"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/vishvananda/netlink/nl"
@@ -24,65 +22,11 @@ import (
 	"github.com/m-lab/tcp-info/tcp"
 )
 
-// Constants from linux.
-const (
-	TCPDIAG_GETSOCK     = 18 // uapi/linux/inet_diag.h
-	SOCK_DIAG_BY_FAMILY = 20 // uapi/linux/sock_diag.h
-)
-
-const TCPF_ALL = 0xFFF
-
-var (
-	// ErrBadPid is used when the PID is mismatched between the netlink socket and the calling process.
-	ErrBadPid = errors.New("bad PID, can't listen to NL socket")
-
-	// ErrBadSequence is used when the Netlink response has a bad sequence number.
-	ErrBadSequence = errors.New("bad sequence number, can't interpret NetLink response")
-
-	// ErrBadMsgData is used when the NHetlink response has bad or missing data.
-	ErrBadMsgData = errors.New("bad message data from netlink message")
-)
-
-// InetDiagReqV2 is the Netlink request struct, as in linux/inet_diag.h
-// Note that netlink messages use host byte ordering, unless NLA_F_NET_BYTEORDER flag is present.
-type InetDiagReqV2 struct {
-	SDiagFamily   uint8
-	SDiagProtocol uint8
-	IDiagExt      uint8
-	Pad           uint8
-	IDiagStates   uint32
-	ID            inetdiag.InetDiagSockID
-}
-
-// SizeofInetDiagReqV2 is the size of the struct.
-// TODO should we just make this explicit in the code?
-const SizeofInetDiagReqV2 = int(unsafe.Sizeof(InetDiagReqV2{})) // Should be 0x38
-
-// Serialize is provided for json serialization?
-// TODO - should use binary functions instead?
-func (req *InetDiagReqV2) Serialize() []byte {
-	return (*(*[SizeofInetDiagReqV2]byte)(unsafe.Pointer(req)))[:]
-}
-
-// Len is provided for json serialization?
-func (req *InetDiagReqV2) Len() int {
-	return SizeofInetDiagReqV2
-}
-
-// NewInetDiagReqV2 creates a new request.
-func NewInetDiagReqV2(family, protocol uint8, states uint32) *InetDiagReqV2 {
-	return &InetDiagReqV2{
-		SDiagFamily:   family,
-		SDiagProtocol: protocol,
-		IDiagStates:   states,
-	}
-}
-
 // TODO - Figure out why we aren't seeing INET_DIAG_DCTCPINFO or INET_DIAG_BBRINFO messages.
 func makeReq(inetType uint8) *nl.NetlinkRequest {
-	req := nl.NewNetlinkRequest(SOCK_DIAG_BY_FAMILY, syscall.NLM_F_DUMP|syscall.NLM_F_REQUEST)
-	msg := NewInetDiagReqV2(inetType, syscall.IPPROTO_TCP,
-		TCPF_ALL & ^((1<<uint(tcp.SYN_RECV))|(1<<uint(tcp.TIME_WAIT))|(1<<uint(tcp.CLOSE))))
+	req := nl.NewNetlinkRequest(inetdiag.SOCK_DIAG_BY_FAMILY, syscall.NLM_F_DUMP|syscall.NLM_F_REQUEST)
+	msg := inetdiag.NewInetDiagReqV2(inetType, syscall.IPPROTO_TCP,
+		inetdiag.TCPF_ALL & ^((1<<uint(tcp.SYN_RECV))|(1<<uint(tcp.TIME_WAIT))|(1<<uint(tcp.CLOSE))))
 	msg.IDiagExt |= (1 << (inetdiag.INET_DIAG_MEMINFO - 1))
 	msg.IDiagExt |= (1 << (inetdiag.INET_DIAG_INFO - 1))
 	msg.IDiagExt |= (1 << (inetdiag.INET_DIAG_VEGASINFO - 1))
@@ -94,7 +38,7 @@ func makeReq(inetType uint8) *nl.NetlinkRequest {
 	msg.IDiagExt |= (1 << (inetdiag.INET_DIAG_SHUTDOWN - 1))
 
 	req.AddData(msg)
-	req.NlMsghdr.Type = SOCK_DIAG_BY_FAMILY
+	req.NlMsghdr.Type = inetdiag.SOCK_DIAG_BY_FAMILY
 	req.NlMsghdr.Flags |= syscall.NLM_F_DUMP | syscall.NLM_F_REQUEST
 	return req
 }
@@ -103,12 +47,12 @@ func processSingleMessage(m *syscall.NetlinkMessage, seq uint32, pid uint32) (*s
 	if m.Header.Seq != seq {
 		log.Printf("Wrong Seq nr %d, expected %d", m.Header.Seq, seq)
 		metrics.ErrorCount.With(prometheus.Labels{"type": "wrong seq num"}).Inc()
-		return nil, false, ErrBadSequence
+		return nil, false, inetdiag.ErrBadSequence
 	}
 	if m.Header.Pid != pid {
 		log.Printf("Wrong pid %d, expected %d", m.Header.Pid, pid)
 		metrics.ErrorCount.With(prometheus.Labels{"type": "wrong pid"}).Inc()
-		return nil, false, ErrBadPid
+		return nil, false, inetdiag.ErrBadPid
 	}
 	if m.Header.Type == unix.NLMSG_DONE {
 		return nil, false, nil
@@ -116,7 +60,7 @@ func processSingleMessage(m *syscall.NetlinkMessage, seq uint32, pid uint32) (*s
 	if m.Header.Type == unix.NLMSG_ERROR {
 		native := nl.NativeEndian()
 		if len(m.Data) < 4 {
-			return nil, false, ErrBadMsgData
+			return nil, false, inetdiag.ErrBadMsgData
 		}
 		error := int32(native.Uint32(m.Data[0:4]))
 		if error == 0 {
