+++ inetdiag/inetdiag_test.go
@@ -1,6 +1,7 @@
 package inetdiag_test
 
 import (
+	"syscall"
 	"testing"
 	"unsafe"
 
@@ -15,5 +16,13 @@ func TestSizes(t *testing.T) {
 	hdr := inetdiag.InetDiagMsg{}
 	if unsafe.Sizeof(hdr) != 4*6+unsafe.Sizeof(inetdiag.InetDiagSockID{}) {
 		t.Error("Header is wrong size", unsafe.Sizeof(hdr))
+	}
+}
+
+func TestInetDiagReqV2Serialize(t *testing.T) {
+	v2 := inetdiag.NewInetDiagReqV2(syscall.AF_INET, 23, 0x0E)
+	data := v2.Serialize()
+	if v2.Len() != len(data) {
+		t.Error(data, "should be length", v2.Len())
 	}
 }
