+++ inetdiag/structs.go
@@ -2,9 +2,65 @@ package inetdiag
 
 import (
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"net"
+	"unsafe"
 )
+
+// Constants from linux.
+const (
+	TCPDIAG_GETSOCK     = 18 // uapi/linux/inet_diag.h
+	SOCK_DIAG_BY_FAMILY = 20 // uapi/linux/sock_diag.h
+)
+
+const TCPF_ALL = 0xFFF
+
+var (
+	// ErrBadPid is used when the PID is mismatched between the netlink socket and the calling process.
+	ErrBadPid = errors.New("bad PID, can't listen to NL socket")
+
+	// ErrBadSequence is used when the Netlink response has a bad sequence number.
+	ErrBadSequence = errors.New("bad sequence number, can't interpret NetLink response")
+
+	// ErrBadMsgData is used when the NHetlink response has bad or missing data.
+	ErrBadMsgData = errors.New("bad message data from netlink message")
+)
+
+// InetDiagReqV2 is the Netlink request struct, as in linux/inet_diag.h
+// Note that netlink messages use host byte ordering, unless NLA_F_NET_BYTEORDER flag is present.
+type InetDiagReqV2 struct {
+	SDiagFamily   uint8
+	SDiagProtocol uint8
+	IDiagExt      uint8
+	Pad           uint8
+	IDiagStates   uint32
+	ID            InetDiagSockID
+}
+
+// SizeofInetDiagReqV2 is the size of the struct.
+// TODO should we just make this explicit in the code?
+const SizeofInetDiagReqV2 = int(unsafe.Sizeof(InetDiagReqV2{})) // Should be 0x38
+
+// Serialize is provided for json serialization?
+// TODO - should use binary functions instead?
+func (req *InetDiagReqV2) Serialize() []byte {
+	return (*(*[SizeofInetDiagReqV2]byte)(unsafe.Pointer(req)))[:]
+}
+
+// Len is provided for json serialization?
+func (req *InetDiagReqV2) Len() int {
+	return SizeofInetDiagReqV2
+}
+
+// NewInetDiagReqV2 creates a new request.
+func NewInetDiagReqV2(family, protocol uint8, states uint32) *InetDiagReqV2 {
+	return &InetDiagReqV2{
+		SDiagFamily:   family,
+		SDiagProtocol: protocol,
+		IDiagStates:   states,
+	}
+}
 
 // InetDiagSockID is the binary linux representation of a socket, as in linux/inet_diag.h
 // Linux code comments indicate this struct uses the network byte order!!!
